+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## [0.5.0] - 2025-05-13
+
+### Added
+- `TreeConditional` node for support of `if-then-else` conditionals in `treelang`.
+
+### Changed
+- Added conditionals support in `AST` parsing, evaluating, visiting and representing (`repr()`)
+- Added conditional node tests.
+- Added conditional example to `Evaluator`.
+- Added example of query with conditional in `calculator.ipynb` cookbook.
+- Added `greater_than` tool in `calculator.py`.
+
+### Fixed
+- Parameter binding for tool creation when multiple arguments have the same name in `AST.tool()`.
+
 ## [0.4.1] - 2025-05-08
 
 ### Fixed